- fixed MTRR mask write bug and enabled check in msr.cc

+++ bios/rombios32.c
@@ -580,7 +580,7 @@ void setup_mtrr(void)
     /* Mark 3.5-4GB as UC, anything not specified defaults to WB */
     wrmsr_smp(MTRRphysBase_MSR(0), 0xe0000000 | MTRR_MEMTYPE_UC);
     /* Make sure no reserved bit set to '1 in MTRRphysMask_MSR */
-    wrmsr_smp(MTRRphysMask_MSR(0), ~(0x20000000 - 1) | 0x800);
+    wrmsr_smp(MTRRphysMask_MSR(0), (uint32_t)(~(0x20000000 - 1)) | 0x800);
     wrmsr_smp(MSR_MTRRdefType, 0xc00 | MTRR_MEMTYPE_UC);
 }
